Display the table with different groups of keywords

Display the result of network-based keyword clustering, with frequency information attached.

keyword_table(tibble_graph, top = 10)

Arguments

tibble_graph
A tbl_graph output by keyword_group.

top
How many keywords should be displayed in the table for each group. Default uses 10.If there is a tie,more than top keywords would be selected. To show all the keywords, use Inf.

Value

A tibble with two columns, namely group and keywords with frequency attached. Different keywords are separated by semicolon(';').

See also

Examples

library(akc)

bibli_data_table %>%
  keyword_clean(id = "id",keyword = "keyword") %>%
  keyword_group(id = "id",keyword = "keyword") %>%
  keyword_table()

#> # A tibble: 7 x 2
#>   Group `Keywords(TOP 10)`                                                      
#>   <int> <chr>                                                                   
#> 1     1 academic libraries (133); library (19); metadata (15); collection manag~
#> 2     2 library services (26); librarians (22); library instruction (21); libra~
#> 3     3 libraries (47); open access (32); bibliometrics (31); library and infor~
#> 4     4 information literacy (58); public libraries (52); university libraries ~
#> 5     5 public library (22); digital library (10); linked data (9); user studie~
#> 6     6 digital humanities (14); open science (10); information retrieval (7); ~
#> 7     7 knowledge management (12); information needs (6); information technolog~