Automatic keyword cleaning and transfer to tidy format

Carry out several keyword cleaning processes automatically and return a tidy table with document ID and keywords.

keyword_clean(
  df,
  id = "id",
  keyword = "keyword",
  sep = ";",
  rmParentheses = TRUE,
  rmNumber = TRUE,
  lemmatize = FALSE
)

Arguments

df
A data.frame containing at least two columns with document ID and keyword strings with separators.

id
Quoted characters specifying the column name of document ID.Default uses "id".

keyword
Quoted characters specifying the column name of keywords.Default uses "keyword".

sep
Separator(s) of keywords. Default uses ";".

rmParentheses
Remove the contents in the parentheses (including the parentheses) or not. Default uses TRUE.

rmNumber
Remove the pure number sequence or no. Default uses TRUE.

lemmatize
Lemmatize the keywords or not. Lemmatization is supported by `lemmatize_strings` function in `textstem` package.Default uses FALSE.

Value

A tbl with two columns, namely document ID and cleaned keywords.

Details

The entire cleaning processes include: 1.Split the text with separators; 2.Reomve the contents in the parentheses (including the parentheses); 3.Remove whitespaces from start and end of string and reduces repeated whitespaces inside a string; 4.Remove all the null character string and pure number sequences; 5.Convert all letters to lower case; 6.Lemmatization. Some of the procedures could be suppressed or activated with parameter adjustments. Default setting did not use lemmatization, it is suggested to use keyword_merge to merge the keywords afterward.

See also

Examples

library(akc)

bibli_data_table

#> # A tibble: 1,448 x 4
#>       id title                  keyword                 abstract                
#>    <int> <chr>                  <chr>                   <chr>                   
#>  1     1 Keeping the doors ope~ Austerity; community c~ "English public librari~
#>  2     2 Comparison of Sloveni~ Comparative librarians~ "This paper aims to pro~
#>  3     3 Analysis of the facto~ Continuation will of v~ "This study aims to dev~
#>  4     4 Redefining Library an~ Curriculum; education ~ "The purpose of this st~
#>  5     5 Can in-house use data~ Check-out use; circula~ "Libraries worldwide ar~
#>  6     6 Practices of communit~ Community councillors;~ "The purpose of the res~
#>  7     7 Exploring Becoming, D~ Library and Informatio~ "Professional identity ~
#>  8     8 Predictors of burnout~ Emotional exhaustion; ~ "Work stress and profes~
#>  9     9 The Roma and document~ Academic libraries; co~ "This paper explores th~
#> 10    10 Mediation effect of k~ Job performance; knowl~ "This paper proposes a ~
#> # ... with 1,438 more rows


bibli_data_table %>%
  keyword_clean(id = "id",keyword = "keyword")

#> # A tibble: 5,378 x 2
#>       id keyword                          
#>    <int> <chr>                            
#>  1     1 austerity                        
#>  2     1 community capacity               
#>  3     1 library professional             
#>  4     1 public libraries                 
#>  5     1 public service delivery          
#>  6     1 volunteer relationship management
#>  7     1 volunteering                     
#>  8     2 comparative librarianship        
#>  9     2 korea                            
#> 10     2 library legislation              
#> # ... with 5,368 more rows